requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
django-filter